% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2GM2D.R
\name{catGenes}
\alias{catGenes}
\title{catGenes}
\usage{
catGenes(geneList, GOGOA3, ontology)
}
\arguments{
\item{geneList}{character vector of user-supplied genes of interest}

\item{GOGOA3}{return value of subsetGOGOA()}

\item{ontology}{c("molecular_function","cellular_component","biological_process")}
}
\value{
returns a matrix of 1's and 0's indicating the presence or absence of gene-category pairs
}
\description{
match up genes in gene list with categories in GOGOA3 database
}
\examples{
#load("data/GOGOA3small.RData")
geneList<-cluster52
m1<-catGenes(geneList,GOGOA3small,"biological_process")

}
