% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2GM2D.R
\name{compareGoMinerHTGM2D}
\alias{compareGoMinerHTGM2D}
\title{compareGoMinerHTGM2D}
\usage{
compareGoMinerHTGM2D(subd, mat, l, ontologies)
}
\arguments{
\item{subd}{character string full path name to the output subdirectory}

\item{mat}{return value of Jaccard()}

\item{l}{of return values of GoMiner()}

\item{ontologies}{character vector of 2 ontologies e.g. c("biological_process","cellular_component")}
}
\value{
returns no value, but saves files that list category difference between GoMiner and HTG2D
}
\description{
compare the results of GoMiner and HTGM2D
}
\examples{
ontologies<-c("biological_process","cellular_component")
#load("data/x_l.Rdata")
#load("data/x_mat.Rdata")
subd<-tempdir()
compareGoMinerHTGM2D(subd,x_mat,x_l,ontologies)

}
