% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2GM2D.R
\name{Jaccard}
\alias{Jaccard}
\title{Jaccard}
\usage{
Jaccard(dir, m1, m2, thresh1 = 2, thresh2 = 3, B = 100)
}
\arguments{
\item{dir}{character string full pathname to the directory acting as result repository}

\item{m1}{return value of catGenes}

\item{m2}{return value of catGenes}

\item{thresh1}{integer acceptance threshold for the number of genes in a cat}

\item{thresh2}{integer acceptance threshold for the number of common genes in 2 cats}

\item{B}{integer a total bootstrap iteration}
}
\value{
returns a numerical matrix containing number of genes and associated p value in the intersection of 2 categories
}
\description{
create the heat map data that is needed as input to JaccardHeatMap()
}
\examples{
#load("data/x_m1.RData")
#load("data/x_m2.RData")
mat<-Jaccard(dir=tempdir(),x_m1,x_m2)

}
