% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2GM2D.R
\name{JaccardHeatMap}
\alias{JaccardHeatMap}
\title{JaccardHeatMap}
\usage{
JaccardHeatMap(dir, mat)
}
\arguments{
\item{dir}{character string containing path name of output directory}

\item{mat}{return value of Jaccard()}
}
\value{
returns a Jaccard matrix of cat1 vs cat2 FDR, and also saves hyperlinked SVG heatmap files to a results directory
}
\description{
use the Jaccard metric to construct 2D heat map
}
\examples{
#load("data/x_jmat.RData")
dir<-tempdir()
jHeatMap<-JaccardHeatMap(dir,x_jmat)

}
